Tidy up for consistency and naming.

+++ esphome/components/partition/light_partition.h
@@ -10,24 +10,24 @@ namespace esphome::partition {
 
 class AddressableSegment {
  public:
-  AddressableSegment(light::LightState *src, int32_t src_offset, int32_t size, bool reversed)
+  AddressableSegment(light::LightState *src, size_t src_offset, size_t size, bool reversed)
       : src_(static_cast<light::AddressableLight *>(src->get_output())),
         src_offset_(src_offset),
         size_(size),
         reversed_(reversed) {}
 
   light::AddressableLight *get_src() const { return this->src_; }
-  int32_t get_src_offset() const { return this->src_offset_; }
-  int32_t get_size() const { return this->size_; }
-  int32_t get_dst_offset() const { return this->dst_offset_; }
-  void set_dst_offset(int32_t dst_offset) { this->dst_offset_ = dst_offset; }
+  size_t get_src_offset() const { return this->src_offset_; }
+  size_t get_size() const { return this->size_; }
+  size_t get_dst_offset() const { return this->dst_offset_; }
+  void set_dst_offset(size_t dst_offset) { this->dst_offset_ = dst_offset; }
   bool is_reversed() const { return this->reversed_; }
 
  protected:
   light::AddressableLight *src_;
-  int32_t src_offset_;
-  int32_t size_;
-  int32_t dst_offset_;
+  size_t src_offset_;
+  size_t size_;
+  size_t dst_offset_;
   bool reversed_;
 };
 
@@ -62,12 +62,12 @@ class PartitionLightOutput final : public light::AddressableLight, protected lig
   }
 
   light::ESPColorView get_color_view(size_t index) override {
-    uint32_t lo = 0;
-    uint32_t hi = this->segments_.size() - 1;
+    size_t lo = 0;
+    size_t hi = this->segments_.size() - 1;
     while (lo < hi) {
-      uint32_t mid = (lo + hi) / 2;
-      int32_t begin = this->segments_[mid].get_dst_offset();
-      int32_t end = begin + this->segments_[mid].get_size();
+      size_t mid = (lo + hi) / 2;
+      size_t begin = this->segments_[mid].get_dst_offset();
+      size_t end = begin + this->segments_[mid].get_size();
       if (index < begin) {
         hi = mid - 1;
       } else if (index >= end) {
@@ -78,9 +78,9 @@ class PartitionLightOutput final : public light::AddressableLight, protected lig
     }
     auto &seg = this->segments_[lo];
     // offset within the segment
-    int32_t seg_off = index - seg.get_dst_offset();
+    size_t seg_off = index - seg.get_dst_offset();
     // offset within the src
-    int32_t src_off;
+    size_t src_off;
     if (seg.is_reversed()) {
       src_off = seg.get_src_offset() + seg.get_size() - seg_off - 1;
     } else {
